+++ packages/expo-image/src/Image.tsx
@@ -132,6 +132,7 @@ export class Image extends React.PureComponent<ImageProps> {
    * @param url - The URL of the image to generate a blurhash from.
    * @param numberOfComponents - The number of components to encode the blurhash with.
    * Must be between 1 and 9. Defaults to `[4, 3]`.
+   * @platform android
    * @platform ios
    * @return A promise resolving to the blurhash string.
    */